[ramda] split tests to separate modules - functions beginning with 'a' (#38157)
* move 'add' tests to separate file

* add README.md with instructions for adding and updating tests

* move 'addIndex' tests to separate file

* move 'all' tests to separate file

* move 'allPass' tests to separate file

* move 'always' tests to separate file

* move 'and' tests to separate file

* move 'any' tests to separate file

* move 'anyPass' tests to separate file

* move 'ap' tests to separate file

* move 'aperture' tests to separate file

* move 'append' tests to separate file

* move 'apply' tests to separate file

* move 'applySpec' tests to separate file

* move 'assoc' tests to separate file

* move 'assocPath' tests to separate file
