+++ types/ramda/test/always-tests.ts
@@ -1,0 +1,6 @@
+import * as R from 'ramda';
+
+() => {
+  const t = R.always('Tee');
+  const x: string = t(); // => 'Tee'
+};
