+++ types/ramda/test/append-tests.ts
@@ -1,0 +1,7 @@
+import * as R from 'ramda';
+
+() => {
+  R.append('tests', ['write', 'more']); // => ['write', 'more', 'tests']
+  R.append('tests')(['write', 'more']); // => ['write', 'more', 'tests']
+  R.append('tests', []); // => ['tests']
+};
